added preliminary file structure to extract logic
