+++ ws_server/src/index.ts
@@ -1,28 +1,17 @@
-import { Request, Response } from "express";
 import { createServer } from "http";
-import { Server, Socket } from "socket.io";
-import { Date, Document } from 'mongoose';
-import { v4 as uuid4 } from 'uuid';
+import { Server } from "socket.io";
+import { Date } from 'mongoose';
+
 const express = require('express');
 const app = express();
 const cors = require('cors');
 const httpServer = createServer(app);
-const { connect } = require("mongoose");
+
 require("dotenv").config();
 
-import dynamoService from "./db/dynamoService";
-const MgRequest = require('./models/request')
 
-// code from Mongoose Typescript Support
-run().catch(err => console.log(err));
-
-// Connect to MongoDB
-async function run() {
-  await connect(process.env.ENV_DB, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  });
-}
+const { sessionIdMiddleware, handleConnection } = require('./services/socketServices')
+const { homeRoute, mongoPostmanRoute, dynamoPostmanRoute } = require('./services/expressServices')
 
 // Express Middleware
 
@@ -44,13 +33,6 @@ interface ServerToClientEvents {
 
 interface RoomData {
   hi: string;
-}
-
-interface IMgRequest extends Document<any> {
-  room: {
-    roomName: string,
-    roomData: RoomData
-  },
 }
 
 interface ClientToServerEvents {
@@ -77,7 +59,7 @@ interface SessionObject {
 
 // instantiating new WS server
 
-const io = new Server<
+export const io = new Server<
   ServerToClientEvents,
   ClientToServerEvents,
   InterServerEvents,
@@ -91,152 +73,16 @@ const io = new Server<
 
 // WS Server Logic
 
-let reconnect: boolean = false;
-
-let currentSessions: SessionObject[] = []
-
-// const fetchMessages = async () => {
-//   let messageArr: IMgRequest[] = await MgRequest.find().sort({ _id: -1 }).limit(5);
-//   return messageArr;
-// }
-
-
-const fetchMissedMessages = async (offset: Date)=> {
-  console.log('#fetchMissedMessages Offset passed', offset)
-  let messageArr: IMgRequest[] = await MgRequest.find({ createdAt: {$gt: offset} });
-  return messageArr;
-}
-
-
-// const fetchLastFive = async (socket: Socket) => {
-//   let messageArr = await fetchMessages();
-//   messageArr.forEach(message => {
-//     let msg = message.room.roomData
-//     socket.emit("connect_message", msg);
-//   });
-// }
-
-
-io.use((socket, next) => {
-  const currentSessionID = socket.handshake.auth.sessionId
-  console.log("Middleware executed");
-  console.log(currentSessionID);
-  console.log(currentSessions);
-
-  // if current session exists (re-connect), find that sessionId from the session
-  if (currentSessionID) {
-    const session = currentSessions.find(obj => obj.sessionId === currentSessionID);
-    if (session) {
-      socket.data.sessionId = session.sessionId;
-      reconnect = true;
-      return next();
-    }
-  }
-  let randomID = uuid4();
-
-  socket.data.sessionId = randomID;
-
-  currentSessions.push({ sessionId: randomID });
-
-  next();
-});
-
-io.on('connection', async (socket) => {
-  if (reconnect) {
-    console.log('A user re-connected');
-    
-    socket.join("room 1");
-    
-    let messageArr = await fetchMissedMessages(socket.handshake.auth.offset)
-
-    messageArr.forEach(message => {
-      let msg = message.room.roomData
-      socket.emit("connect_message", msg);
-    });
-
-//     fetchLastFive(socket);
-
-  } else {
-    console.log('A user connected first time');
-    socket.join("room 1");
-    socket.handshake.auth.offset = undefined
-    socket.emit("session", {
-      sessionId: socket.data.sessionId,
-    })
-  }
-
-//   socket.on("disconnecting", (reason) => {
-//     if (reason === "client namespace disconnect") {
-//       reconnect = true;
-//       // push an object with session_id and unintentionalDisconnect
-//     }
-//   });
-  
-  socket.on('disconnect', () => {
-    console.log('user disconnected');
-  });
-});
-
-/// atLeastOnce server-side START ////////
-
-// io.on("connection", async (socket) => {
-//   const offset = socket.handshake.auth.offset;
-//   if (offset) {
-//     // this is a reconnection
-//     for (const event of await fetchMissedEventsFromDatabase(offset)) {
-//       socket.emit("my-event", event);
-//     }
-//   } else {
-//     // this is a first connection
-//   }
-// });
-
-/// atLeastOnce server-side END ////////
+io.use(sessionIdMiddleware)
+io.on("connection", handleConnection)
 
 // Backend API
 
-app.get('/', (req: Request, res: Response) => {
-  console.log("you've got mail!");
-  res.send('Nice work')
-});
+app.get('/', homeRoute);
+app.put('/api/postman', mongoPostmanRoute);
+app.post('/api/postman/dynamo', dynamoPostmanRoute);
 
-app.put('/api/postman', async (req: Request, res: Response) => {
-  // accept postman put request
-  // publish this request.body data via websocket emit
-  const data: string = req.body;
-  console.log(data)
-
-  const currentRequest = new MgRequest({
-    room: {
-      roomName: "room 1",
-      roomData: data
-    },
-  });
-
-  const savedRequest = await currentRequest.save();
-
-  const timestamp: Date = savedRequest.createdAt
-  let messageData: any[] = [data, timestamp]
-  
-  io.to("room 1").emit("message", messageData);
-
-  console.log('SENT POSTMAN MESSAGE');
-
-  res.send('ok');
-});
-
-// need create an interface for the request body
-app.post('/api/postman/dynamo', async (req: Request, res: Response) => {
-  try {
-    const data: any = req.body;  // specify the actual type
-    const dynamoResponse: any = await dynamoService.createMessage(data.RoomId, data.Message) // specify the actual type
-    console.log('SENT POSTMAN MESSAGE:', data.Message);
-    io.to("room 1").emit("message", data.Message);
-    res.status(dynamoResponse['$metadata']['httpStatusCode']).send('ok');
-  } catch (error) {
-    console.log(error);
-  }
-});
+// listening on port 3001
 
 httpServer.listen(PORT, () => {
   console.log('listening on port', PORT);
